requirements.txt
Django==4.2.23
djangorestframework==3.16.1
django-cors-headers==4.7.0
asgiref==3.9.1
sqlparse==0.5.3
typing-extensions==4.14.1
requests==2.32.4
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
PyJWT==2.10.1
cryptography==45.0.6
cffi==1.17.1
pycparser==2.22
dj-rest-auth==7.0.1
python-decouple
djangorestframework-simplejwt
openai
mysqlclient
